Move optimization/SKOpt assets to new `optimization` directory
- This mostly affects three core modules: `optimization_core`,
  `optimization`, and `space`
- The above three modules have been moved to a new `optimization`
  directory
   - `optimization_core` -> `optimization.protocol_core`
   - `optimization` -> `optimization.backends.skopt.protocols`
   - `space` -> `optimization.backends.skopt.space`
- In `optimization.backends.skopt`, a copy of SKOpt's license, and a
  readme file have been added to clarify that the contents of the
  module are adapted SKOpt utilities
- All other code changes outside of the three aforementioned modules
  are import statement corrections for the restructuring

- This change was made both to consolidate optimization/space assets
  in a single module and to pave the way for adding more
  `optimization.backends` providers
- Additionally, it was necessary to add SKOpt's license, especially as
  further modifications are required
