+++ hyperparameter_hunter/__init__.py
@@ -1,4 +1,15 @@
 ##################################################
+# Store Library Version
+##################################################
+import os.path
+
+try:
+    with open(os.path.join(os.path.dirname(__file__), "VERSION")) as f:
+        __version__ = f.read().strip()
+except Exception:
+    raise
+
+##################################################
 # Execute Import Interceptors
 ##################################################
 from pkg_resources import DistributionNotFound
@@ -13,39 +24,28 @@ except Exception:
     raise
 
 ##################################################
-# Store Library Version
-##################################################
-import os.path
-
-try:
-    with open(os.path.join(os.path.dirname(__file__), "VERSION")) as f:
-        __version__ = f.read().strip()
-except Exception:
-    raise
-
-##################################################
 # Set __all__
 ##################################################
 from .environment import Environment
 from .experiments import CVExperiment
-from .optimization import BayesianOptPro
-from .optimization import GradientBoostedRegressionTreeOptPro, GBRT
-from .optimization import RandomForestOptPro, RF
-from .optimization import ExtraTreesOptPro, ET
-from .optimization import DummyOptPro
-from .space import Real
-from .space import Integer
-from .space import Categorical
+from .optimization.backends.skopt.protocols import BayesianOptPro
+from .optimization.backends.skopt.protocols import GradientBoostedRegressionTreeOptPro, GBRT
+from .optimization.backends.skopt.protocols import RandomForestOptPro, RF
+from .optimization.backends.skopt.protocols import ExtraTreesOptPro, ET
+from .optimization.backends.skopt.protocols import DummyOptPro
+from .optimization.backends.skopt.space import Real
+from .optimization.backends.skopt.space import Integer
+from .optimization.backends.skopt.space import Categorical
 from .callbacks.bases import lambda_callback
 from .feature_engineering import FeatureEngineer
 from .feature_engineering import EngineerStep
 
 #################### Deprecated OptPros - Remove in 3.2.0 ####################
-from .optimization import BayesianOptimization
-from .optimization import GradientBoostedRegressionTreeOptimization
-from .optimization import RandomForestOptimization
-from .optimization import ExtraTreesOptimization
-from .optimization import DummySearch
+from .optimization.backends.skopt.protocols import BayesianOptimization
+from .optimization.backends.skopt.protocols import GradientBoostedRegressionTreeOptimization
+from .optimization.backends.skopt.protocols import RandomForestOptimization
+from .optimization.backends.skopt.protocols import ExtraTreesOptimization
+from .optimization.backends.skopt.protocols import DummySearch
 
 __all__ = [
     #################### Environment ####################
