+++ hyperparameter_hunter/optimization/backends/skopt/README.md
@@ -1,0 +1,9 @@
+The contents of this module are based on utilities from the excellent [Scikit-Optimize](https://github.com/scikit-optimize/scikit-optimize)
+library. Therefore, any code contained in this module originally from SKOpt is licensed under
+the "New BSD License" (3-clause BSD license), which can be found in this module, [here](LICENSE). 
+
+Any additions or modifications not originating from Scikit-Optimize fall under 
+HyperparameterHunter's license, [here](../../../../LICENSE).
+
+Thank you to the Scikit-Optimize developers for their fantastic work and for allowing other to 
+build on that work.
