+++ hyperparameter_hunter/optimization/backends/skopt/protocols.py
@@ -11,8 +11,8 @@ Related
 # Import Own Assets
 ##################################################
 from hyperparameter_hunter import __version__
-from hyperparameter_hunter.optimization_core import SKOptPro
-from hyperparameter_hunter.space import normalize_dimensions
+from hyperparameter_hunter.optimization.protocol_core import SKOptPro
+from hyperparameter_hunter.optimization.backends.skopt.space import normalize_dimensions
 from hyperparameter_hunter.utils.version_utils import Deprecated
 
 ##################################################
