+++ hyperparameter_hunter/reporting.py
@@ -2,8 +2,8 @@
 # Import Own Assets
 ##################################################
 from hyperparameter_hunter import exceptions
+from hyperparameter_hunter.optimization.backends.skopt.space import Categorical
 from hyperparameter_hunter.settings import G
-from hyperparameter_hunter.space import Categorical
 from hyperparameter_hunter.utils.general_utils import now_time, expand_mins_secs
 
 ##################################################
